Add experimental VS Code (`code`) runtime to `juv run`
This PR adds a `code` runtime option to `juv run`, allowing VS Code to
be launched in a way similar to other Jupyter runtimes. The approach
uses the behavior of the `code` CLI, which inherits the environment
it was run in.

A script is created using the inline metadata so that running `code
<notebook.ipynb>` ensures a virtual environment is available in the VS
Code session with the necessary dependencies.

Ideally, we'd have a VS Code extension to dynamically create and
register virtual environments for better discoverability. This
implementation goes the other way by enabling basic support via the CLI.
However, the user must manually select the virtual environment from VS
Code's dropdown, where it appears with an obscure, temporary name
(derived from the hashed environment created by `uv`).